requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
slowapi==0.1.9
python-multipart==0.0.9
argon2-cffi==23.1.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7